Create `HelmApp`
